# internal/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Event Flow API
  version: 0.1.0
  description: |
    Сервис регистрации на мероприятия.

    Как пользоваться в Swagger UI:
    1. `POST /auth/register` — создать аккаунт (роль `organizer` или `participant`).
    2. `POST /auth/login` — получить `token`.
    3. Нажать **Authorize** и вставить токен (без слова `Bearer`).

    Пользователь всегда определяется по токену сессии, а не по полям запроса (D-01).
    Для чужих и несуществующих объектов ответ одинаковый — 404 (D-02).
servers:
  - url: /
tags:
  - name: system
    description: Служебные эндпоинты
  - name: auth
    description: Аккаунт и сессии
  - name: profile
    description: Профиль текущего пользователя
  - name: events
    description: Мероприятия
  - name: registrations
    description: Записи на мероприятия

paths:
  /health:
    get:
      tags: [system]
      operationId: health
      security: []
      summary: Проверка, что сервер и БД работают
      responses:
        "200":
          description: Всё работает
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
              example: { status: ok, db: ok }
        "503":
          description: БД недоступна
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
              example: { status: error, db: unavailable }

  /auth/register:
    post:
      tags: [auth]
      operationId: registerAccount
      security: []
      summary: Создать аккаунт
      description: Роль выбирается один раз и дальше через API не меняется (SR-08).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Аккаунт создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Логин уже занят
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      summary: Войти и получить токен сессии
      description: После 5 неверных паролей подряд вход блокируется на 15 минут (SR-10).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Неверный логин или пароль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Аккаунт временно заблокирован после неудачных попыток
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Выйти (токен сразу перестаёт работать)
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Сессия удалена }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /me:
    get:
      tags: [profile]
      operationId: getMe
      summary: Мой профиль
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [profile]
      operationId: updateMe
      summary: Изменить имя и контакты
      description: Можно передать только `name` и `contacts`. Любое другое поле (например, `role`) — 400.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchMeRequest" }
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /events:
    get:
      tags: [events]
      operationId: listEvents
      security: []
      summary: Список опубликованных мероприятий
      description: Доступно без входа. Из данных организатора отдаётся только имя (SR-11).
      responses:
        "200":
          description: Список мероприятий
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
    post:
      tags: [events]
      operationId: createEvent
      summary: Опубликовать мероприятие (только организатор)
      description: Организатором становится текущий пользователь.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEventRequest" }
      responses:
        "201":
          description: Мероприятие опубликовано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  /events/{id}:
    get:
      tags: [events]
      operationId: getEvent
      security: []
      summary: Одно мероприятие
      parameters: [{ $ref: "#/components/parameters/EventID" }]
      responses:
        "200":
          description: Мероприятие
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "404": { $ref: "#/components/responses/NotFound" }

  /events/{id}/participants:
    get:
      tags: [events]
      operationId: listParticipants
      summary: Участники моего мероприятия (только его организатор)
      description: Для чужого мероприятия — 404 (SR-09).
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/EventID" }]
      responses:
        "200":
          description: Активные участники
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Participant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  /events/{id}/registrations:
    post:
      tags: [registrations]
      operationId: registerForEvent
      summary: Записаться на мероприятие (только участник)
      description: Тело запроса не нужно и не читается. Запись идёт на текущего пользователя.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/EventID" }]
      responses:
        "201":
          description: Запись создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Registration" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Мест нет или пользователь уже записан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                noSeats: { value: { error: no free seats } }
                already: { value: { error: already registered } }

  /registrations/my:
    get:
      tags: [registrations]
      operationId: listMyRegistrations
      summary: Мои записи (включая отменённые)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Список записей
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Registration" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /registrations/{id}:
    delete:
      tags: [registrations]
      operationId: cancelRegistration
      summary: Отменить свою запись
      description: Для чужой, несуществующей или уже отменённой записи — 404. Место возвращается ровно один раз.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        "204": { description: Запись отменена }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен из ответа `POST /auth/login`.

  parameters:
    EventID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }

  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Нет токена, токен неверный или истёк
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: unauthorized }
    Forbidden:
      description: Недостаточно прав для роли
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: forbidden }
    NotFound:
      description: Не найдено (или нет доступа)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: not found }

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }

    Health:
      type: object
      properties:
        status: { type: string, enum: [ok, error] }
        db: { type: string, enum: [ok, unavailable] }

    Role:
      type: string
      enum: [organizer, participant]

    RegisterRequest:
      type: object
      additionalProperties: false
      required: [login, password, name, role]
      properties:
        login: { type: string, minLength: 3, maxLength: 64, example: org }
        password: { type: string, format: password, minLength: 8, maxLength: 128, example: password123 }
        name: { type: string, minLength: 1, maxLength: 100, example: Иван Петров }
        contacts: { type: string, maxLength: 200, example: org@mail.ru }
        role: { $ref: "#/components/schemas/Role" }

    LoginRequest:
      type: object
      additionalProperties: false
      required: [login, password]
      properties:
        login: { type: string, example: org }
        password: { type: string, format: password, example: password123 }

    LoginResponse:
      type: object
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }

    Me:
      type: object
      properties:
        id: { type: integer, format: int64 }
        login: { type: string }
        name: { type: string }
        contacts: { type: string }
        role: { $ref: "#/components/schemas/Role" }

    PatchMeRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        contacts: { type: string, maxLength: 200 }

    CreateEventRequest:
      type: object
      additionalProperties: false
      required: [title, starts_at, capacity]
      properties:
        title: { type: string, minLength: 1, maxLength: 200, example: Go meetup }
        description: { type: string, maxLength: 5000, example: Доклады про Go }
        starts_at: { type: string, format: date-time, description: Время в будущем (RFC 3339), example: "2030-01-01T10:00:00Z" }
        capacity: { type: integer, minimum: 1, maximum: 100000, example: 30 }

    Event:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        description: { type: string }
        starts_at: { type: string, format: date-time }
        capacity: { type: integer }
        free_seats: { type: integer }
        status: { type: string, enum: [published, cancelled] }
        organizer_name: { type: string }

    Participant:
      type: object
      properties:
        name: { type: string }
        contacts: { type: string }
        registered_at: { type: string, format: date-time }

    Registration:
      type: object
      properties:
        id: { type: integer, format: int64 }
        event_id: { type: integer, format: int64 }
        event_title: { type: string }
        event_starts_at: { type: string, format: date-time }
        status: { type: string, enum: [active, cancelled] }
        created_at: { type: string, format: date-time }